added view user and edit user components
