+++ frontend/pages/UserProfileScreen.js
@@ -3,6 +3,8 @@ import { Context } from "../store/context";
 import { useContext } from "react";
 import { useRouter } from "next/router";
 import ManagerScreen from "./ManagerScreen";
+import SystemAdminScreen from "./SystemAdminScreen";
+
 const DisplayUserProfile = () => {
   const [currentUser, setCurrentUser] = useContext(Context);
   const router = useRouter();
@@ -17,6 +19,9 @@ const DisplayUserProfile = () => {
         <title>User Profile</title>
       </Head>
       {currentUser && currentUser.userType === "manager" && <ManagerScreen />}
+      {currentUser && currentUser.userType === "systemAdmin" && (
+        <SystemAdminScreen />
+      )}
       <div className="text-right m-4">
         <button className="m-4 p-2 bg-cyan-900" onClick={logOut}>
           Log Out
